+++ Packages/arch.systemgroups.visualiser/Editor/Windows/SystemsWindow.cs
@@ -1,15 +1,15 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Arch.SystemGroups;
+using SystemGroups.Visualiser.Editor.Tree;
 using SystemGroups.Visualiser.Editor.Utility;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.UIElements;
 
-namespace SystemGroups.Visualiser.Editor
+namespace SystemGroups.Visualiser.Editor.Windows
 {
-    public class SystemsWindow : EditorWindow
+    public partial class SystemsWindow : EditorWindow
     {
         private static readonly string kPackagePath = "Packages/arch.systemgroups.visualiser/Editor/Assets";
         private MultiColumnTreeView _multiColumnTreeView;
@@ -72,7 +72,8 @@ namespace SystemGroups.Visualiser.Editor
             if (_systemFilterMenu != null) _systemFilterMenu.RegisterValueChangedCallback(_filterCallback);
             if (_worldDropdownMenu != null) _worldDropdownMenu.RegisterValueChangedCallback(OnSystemGroupWorldValueChanged);
             SystemGroupSnapshot.Instance.OnSystemGroupWorldChanged += OnSystemGroupWorldChanged;
-            
+
+            _multiColumnTreeView.selectionChanged += OnSelectionChanged;
             _multiColumnTreeView.RegisterCallback<ContextClickEvent>(ExpandMenu);
         }
 
@@ -137,7 +138,8 @@ namespace SystemGroups.Visualiser.Editor
                 _debouncer.Debounce(() =>
                 {
                     PopulateHierarchy();
-                    _multiColumnTreeView.ExpandAll(); 
+                    _multiColumnTreeView.ExpandAll();
+                    _multiColumnTreeView.ScrollItemToTop(_lastSelectionId);
                 });
             };
 
@@ -153,6 +155,8 @@ namespace SystemGroups.Visualiser.Editor
             _worldDropdownMenu = rootVisualElement.Q<DropdownField>("world-dropdown");
             _systemFilterMenu = rootVisualElement.Q<TextField>("systems-filter");
             _enterPlayMode = rootVisualElement.Q<Button>("enter-play-mode");
+
+            _multiColumnTreeView.selectionType = SelectionType.Single;
         }
 
 
@@ -203,99 +207,14 @@ namespace SystemGroups.Visualiser.Editor
             _multiColumnTreeView.SetRootItems(Array.Empty<TreeViewItemData<Descriptor>>());
 
             var rootDescriptor = new List<TreeViewItemData<Descriptor>>();
-            GenerateSystemData(rootDescriptor, _systemFilterMenu.value);
+            SystemTreeBuilder.Generate(rootDescriptor, _systemFilterMenu.value, _worldDropdownMenu.value);
 
             // Setup System/Group Name Column
-
             SetupNameColumn();
             SetupThrottledColumn();
-            
+
             _multiColumnTreeView.SetRootItems(rootDescriptor);
             _multiColumnTreeView.Rebuild();
-        }
-
-        private void SetupThrottledColumn()
-        {
-            const string COLUMN_NAME = "throttled-column";  
-            var throttledColumn = _multiColumnTreeView.columns.FirstOrDefault(c => c.name == COLUMN_NAME);
-            if(throttledColumn == null) return;
-            
-            throttledColumn.makeCell = () => _rowTemplate.CloneTree();
-            throttledColumn.bindCell = (element, rowIndex) =>
-            {
-                var node = _multiColumnTreeView.GetItemDataForIndex<Descriptor>(rowIndex);
-                var icon = element.Q<Image>("icon");
-                var label = element.Q<Label>("name");
-
-                if (node.IsGroup)
-                {
-                    label.text = String.Empty;
-                }
-                else
-                {
-                    label.text = node.ThrottlingEnabled ? "Enabled" : "Disabled";    
-                }
-                
-                icon.EnableInClassList("group", false);
-                icon.EnableInClassList("system", false);
-            };
-        }
-
-        private void SetupNameColumn()
-        {
-            const string COLUMN_NAME = "name";
-            var nameColumn = _multiColumnTreeView.columns.FirstOrDefault(c => c.name == COLUMN_NAME);
-            if (nameColumn == null) return;
-            
-            nameColumn.makeCell = () => _rowTemplate.CloneTree();
-            nameColumn.bindCell = (element, rowIndex) =>
-            {
-                var node = _multiColumnTreeView.GetItemDataForIndex<Descriptor>(rowIndex);
-                var icon = element.Q<Image>("icon");
-                var label = element.Q<Label>("name");
-                
-                label.enableRichText = true;
-                var query = _systemFilterMenu?.value ?? String.Empty;
-                label.text = RichText.HighlightMatch(node.Name, query);
-                
-                icon.EnableInClassList("group", node.IsGroup);
-                icon.EnableInClassList("system", node.IsSystem);
-            };
-        }
-
-        /// <summary>
-        /// GenerateSystemData
-        /// </summary>
-        /// <param name="idx"></param>
-        /// <param name="parent"></param>
-        /// <returns></returns>
-        private void GenerateSystemData(IList<TreeViewItemData<Descriptor>> parent, string filter)
-        {
-            void Recurse(IReadOnlyList<Descriptor> descriptors, IList<TreeViewItemData<Descriptor>> parent)
-            {
-                foreach (var descriptor in descriptors)
-                {
-                    if (descriptor.IsGroup)
-                    {
-                        var children = new List<TreeViewItemData<Descriptor>>();
-                        var tvi = new TreeViewItemData<Descriptor>(descriptor.Name.GetHashCode(), descriptor, children);
-                        parent.Add(tvi);
-                        Recurse(descriptor.SubDescriptors, children);
-                    }
-                    else if(descriptor.IsSystem)
-                    {
-                        if(!String.IsNullOrEmpty(filter) && !descriptor.Name.Contains(filter, StringComparison.OrdinalIgnoreCase)) continue;
-                        var tvi = new TreeViewItemData<Descriptor>(descriptor.Name.GetHashCode(), descriptor);
-                        parent.Add(tvi);                    
-                    }
-                }
-            }
-            
-            var descriptors = SystemGroupSnapshot.Instance.Capture(_worldDropdownMenu.value);
-            if (descriptors != null)
-            {
-                Recurse(descriptors, parent);   
-            }
         }
 
         /// <summary>
@@ -306,6 +225,7 @@ namespace SystemGroups.Visualiser.Editor
             if (_multiColumnTreeView != null)
             {
                 _multiColumnTreeView.SetRootItems(Array.Empty<TreeViewItemData<Descriptor>>());
+                _multiColumnTreeView.Rebuild();
             }
         }
     }
